refactor and partialy fix ios e2e tests
