+++ examples/webdriverio-test/.env.example
@@ -6,3 +6,4 @@ ANDROID_APP_PATH=path/to/your/repo/examples/fishjam-chat/android/app/build/outpu
 IOS_DEVICE_ID=123123123
 IOS_TEAM_ID=123123123
 IOS_APP_PATH=path/to/your/ipa/folder/FishjamExample.ipa
+IOS_TEST_SCREEN_BROADCAST=true
