+++ examples/webdriverio-test/models.ts
@@ -1,0 +1,5 @@
+export type Test = {
+  name: string;
+  run: () => Promise<void>;
+  skip: boolean;
+};
