+++ examples/webdriverio-test/package.json
@@ -12,6 +12,7 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
+    "@fishjam-cloud/fishjam-openapi": "^0.1.2",
     "@wdio/appium-service": "^8.40.2",
     "@wdio/cli": "^8.40.5",
     "@wdio/local-runner": "^8.40.5",
@@ -23,7 +24,6 @@
     "webdriverio": "^9.1.1"
   },
   "dependencies": {
-    "@fishjam-cloud/js-server-sdk": "^0.1.2",
     "dotenv": "^16.3.1"
   }
 }
